Exception handling - delete

+++ src/main/java/com/ufersa/course/services/UserService.java
@@ -4,10 +4,12 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.ufersa.course.entities.User;
 import com.ufersa.course.repositories.UserRepository;
+import com.ufersa.course.services.exceptions.DatabaseException;
 import com.ufersa.course.services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -29,8 +31,15 @@ public class UserService {
 		return repository.save(user);
 	}
 	
-	public void delete(Long id) {
-		repository.deleteById(id);;
+	public void delete(Long id){
+	    try {
+	        if(!repository.existsById(id)) throw new ResourceNotFoundException(id);
+	        repository.deleteById(id);
+	    }catch (ResourceNotFoundException e){
+	        throw new ResourceNotFoundException(id);
+	    }catch (DataIntegrityViolationException e) {
+	    	throw new DatabaseException(e.getMessage());
+	    }
 	}
 	
 	public User update(Long id, User user) {
